:truck: move the static files into the src folder
