+++ src/components/widget/widget.stories.js
@@ -20,7 +20,7 @@ import {
   MenuDivider,
   Bullet,
 } from '../../index';
-import { rows1 } from './static/data/datagrid';
+import { rows1 } from '../../static/data/datagrid';
 import { IconAddMediumOutline, IconEditMediumOutline } from '@teamleader/ui-icons';
 
 const colors = ['mint', 'violet', 'ruby', 'gold', 'aqua', 'white', 'neutral'];
