+++ src/static/data/avatar.js
